Simplify connection initialization
