+++ projectBlueWater/src/main/java/com/lasthopesoftware/bluewater/client/connection/selected/InstantiateSelectedConnectionActivity.kt
@@ -35,8 +35,7 @@ class InstantiateSelectedConnectionActivity : AppCompatActivity(), ControlConnec
 	private val libraryConnectionProvider by lazy { buildNewConnectionSessionManager() }
 
 	private val connectionInitializationProxy by lazy {
-		DramaticConnectionInitializationController(
-			ConnectionInitializationProxy(libraryConnectionProvider),
+		DramaticConnectionInitializationProxy(
 			libraryConnectionProvider,
 		)
 	}
