--- projectBlueWater/src/main/java/com/lasthopesoftware/bluewater/client/connection/session/initialization/ConnectionInitializationProxy.kt
@@ -1,15 +1,0 @@
-package com.lasthopesoftware.bluewater.client.connection.session.initialization
-
-import com.lasthopesoftware.bluewater.client.browsing.library.repository.LibraryId
-import com.lasthopesoftware.bluewater.client.connection.BuildingConnectionStatus
-import com.lasthopesoftware.bluewater.client.connection.IConnectionProvider
-import com.lasthopesoftware.bluewater.client.connection.session.ManageConnectionSessions
-import com.lasthopesoftware.bluewater.shared.promises.extensions.ProgressingPromise
-import com.lasthopesoftware.bluewater.shared.promises.extensions.ProgressingPromiseProxy
-
-class ConnectionInitializationProxy(
-	private val manageConnectionSessions: ManageConnectionSessions,
-) : ControlConnectionInitialization {
-	override fun promiseInitializedConnection(libraryId: LibraryId): ProgressingPromise<BuildingConnectionStatus, IConnectionProvider?> =
-		ProgressingPromiseProxy(manageConnectionSessions.promiseLibraryConnection(libraryId))
-}
